fix(hv): distinguish between VCPU index and identity
Alioth uses a VCPU index (0..num_cpus) to manage VCPUs internally.
However, the underlying hypervisor or the hardware uses an
architecture-dependent identity for each CPU, such as MPIDR on aarch64
and APIC ID on x86_64.

This commit introduces a clear distinction between the VCPU index and
the VCPU identity:

- The VCPU index is consistently represented as `u16`.
- A new method `encode_cpu_identity(index: u16) -> u64` is added to
  struct `Board` to translate an index into the corresponding identity.
- The `Vm::create_vcpu` method is updated to accept both the index and
  the identity.
- On aarch64, Alioth now generates and provides the MPIDR to the
  hypervisor, ensuring a consistent CPU topology.

+++ alioth-cli/src/boot.rs
@@ -186,7 +186,7 @@ pub struct BootArgs {
 
     /// Number of VCPUs assigned to the guest.
     #[arg(long, default_value_t = 1)]
-    num_cpu: u32,
+    num_cpu: u16,
 
     /// DEPRECATED: Use --memory instead.
     #[arg(long, default_value = "1G")]
